requirements.txt
opencv-python
pyyaml
ultralytics
streamlit
pillow